Step 2: Interface + Mock
At this step I used Copilot Edits to extract the interface from the codeunit, apply dependency injection, and then create a mock for the interface.

The prompts for both steps are provided in the slide deck of the session.

Only minimal edits were made (to the interface file) to eliminate problems that Copilot has with understanding what AL can (and cannot) do (where it still struggles).