Step 4: Completing code
Here the code is completed.

Copilot has created all interfaces, wrappers, and their corresponding mocks (with the prompts provided in the slide deck). Also, Copilot was used to keep the `ProcessQuotesMock` codeunit up-to-date with changes.

Also, Copilot has correctly guessed most of the places (in functional code) where wrappers are used and  dependency injection was needed.